feat: verify ecdsa on hathorlib
